Start refactor to allow compaction to multi file too

+++ xapian-core/backends/glass/glass_compact.cc
@@ -932,25 +932,31 @@ compact_glass(Xapian::Compactor & compactor,
 	version_file[i].read();
     }
 
-    rmdir(destdir);
-    int fd = open(destdir, O_RDWR|O_CREAT|O_BINARY|O_CLOEXEC, 0666);
-    if (fd < 0) {
-	throw Xapian::DatabaseCreateError("open() failed", errno);
-    }
-#if 0
-    FlintLock lock(destdir);
-    string explanation;
-    FlintLock::reason why = lock.lock(true, false, explanation);
-    if (why != FlintLock::SUCCESS) {
-	lock.throw_databaselockerror(why, destdir, explanation);
-    }
-#endif
+    bool single_file = true;
 
-    GlassVersion version_file_out(fd); //destdir);
+    FlintLock lock(destdir);
+    if (!single_file) {
+	string explanation;
+	FlintLock::reason why = lock.lock(true, false, explanation);
+	if (why != FlintLock::SUCCESS) {
+	    lock.throw_databaselockerror(why, destdir, explanation);
+	}
+    }
+
+    GlassVersion version_file_out(destdir);
+    int fd = -1;
+    if (single_file) {
+	rmdir(destdir);
+	fd = open(destdir, O_RDWR|O_CREAT|O_BINARY|O_CLOEXEC, 0666);
+	if (fd < 0) {
+	    throw Xapian::DatabaseCreateError("open() failed", errno);
+	}
+	version_file_out.set_fd(fd);
+    }
     version_file_out.create(block_size, 0);
 
     string fl_serialised;
-    {
+    if (single_file) {
 	GlassFreeList fl;
 	fl.set_first_unused_block(1); // FIXME: Assumption?
 	fl.pack(fl_serialised);
@@ -967,12 +973,10 @@ compact_glass(Xapian::Compactor & compactor,
 	// from each source table in turn.
 	compactor.set_status(t->name, string());
 
-	/*
 	string dest = destdir;
 	dest += '/';
 	dest += t->name;
 	dest += '.';
-	*/
 
 	bool output_will_exist = !t->lazy;
 
@@ -1023,16 +1027,20 @@ compact_glass(Xapian::Compactor & compactor,
 	    continue;
 	}
 
-	GlassTable * out =
-	    new GlassTable(t->name, dup(fd) /*dest*/, false, t->compress_strategy, false /*t->lazy*/);
+	GlassTable * out;
+	if (single_file) {
+	    out = new GlassTable(t->name, fd, false, t->compress_strategy, false);
+	} else {
+	    out = new GlassTable(t->name, dest, false, t->compress_strategy, t->lazy);
+	}
 	tabs.push_back(out);
 	RootInfo * root_info;
-	if (false) {
-	    out->create_and_open(FLAGS, block_size);
-	} else {
+	if (single_file) {
 	    root_info = version_file_out.root_to_set(t->type);
 	    root_info->set_free_list(fl_serialised);
 	    out->open(FLAGS, version_file_out.get_root(t->type), version_file_out.get_revision());
+	} else {
+	    out->create_and_open(FLAGS, block_size);
 	}
 
 	out->set_full_compaction(compaction != compactor.STANDARD);
@@ -1077,18 +1085,27 @@ compact_glass(Xapian::Compactor & compactor,
 	}
 
 	// Commit as revision 1.
-	lseek(fd, 0, SEEK_SET);
+	if (single_file) lseek(fd, 0, SEEK_SET);
 	out->flush_db();
 	out->commit(1, version_file_out.root_to_set(t->type));
 	out->sync();
-	fl_serialised = root_info->get_free_list();
+	if (single_file) fl_serialised = root_info->get_free_list();
 
 	off_t out_size = 0;
 	if (!bad_stat) {
-	    off_t db_size = file_size(fd); // dest + GLASS_TABLE_EXTENSION);
+	    off_t db_size;
+	    if (single_file) {
+		db_size = file_size(fd);
+	    } else {
+		db_size = file_size(dest + GLASS_TABLE_EXTENSION);
+	    }
 	    if (errno == 0) {
-		out_size = (db_size - prev_size) / 1024;
-		prev_size = db_size;
+		if (single_file) {
+		    off_t old_prev_size = prev_size;
+		    prev_size = db_size;
+		    db_size -= old_prev_size;
+		}
+		out_size = db_size / 1024;
 	    } else {
 		bad_stat = (errno != ENOENT);
 	    }
@@ -1133,7 +1150,5 @@ compact_glass(Xapian::Compactor & compactor,
 	delete tabs[j];
     }
 
-#if 0
-    lock.release();
-#endif
+    if (!single_file) lock.release();
 }
